+++ lib/app/homePages.dart
@@ -4,6 +4,7 @@ import 'package:logger/logger.dart';
 
 import 'package:march_tales_app/core/types/HomePageData.dart';
 import 'package:march_tales_app/pages/FavoriteTracksPage.dart';
+import 'package:march_tales_app/pages/RecentsPage.dart';
 import 'package:march_tales_app/pages/SettingsPage.dart';
 import 'package:march_tales_app/pages/TracksPage.dart';
 import 'homePages.i18n.dart';
@@ -11,13 +12,19 @@ import 'homePages.i18n.dart';
 final logger = Logger();
 
 enum HomePages {
-  root,
+  recents,
+  tracks,
   favorites,
   settings,
 }
 
 List<HomePageData> getHomePages() {
   final homePages = [
+    HomePageData(
+      widget: () => RecentsPage(),
+      label: 'Home'.i18n,
+      icon: Icon(Icons.home),
+    ),
     HomePageData(
       widget: () => TracksPage(),
       label: 'Tracks'.i18n,