+++ lib/pages/FavoriteTracksPage.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:logger/logger.dart';
 import 'package:provider/provider.dart';
 
-import 'package:march_tales_app/app/AppColors.dart';
 import 'package:march_tales_app/blocks/NoTracksInfo.dart';
+import 'package:march_tales_app/components/LoadingSplash.dart';
 import 'package:march_tales_app/features/Track/widgets/TracksList.dart';
 import 'package:march_tales_app/shared/states/AppState.dart';
 
@@ -46,13 +46,13 @@ class FavoriteTracksPage extends StatelessWidget {
     final isFavoritesLoading = appState.isFavoritesLoading;
     final tracks = appState.getSortedFavorites();
 
-    final theme = Theme.of(context);
-    final AppColors appColors = theme.extension<AppColors>()!;
+    // final theme = Theme.of(context);
+    // final AppColors appColors = theme.extension<AppColors>()!;
 
     final isLoading = !favoritesHasBeenLoaded && isFavoritesLoading;
     if (isLoading) {
       return ShowWrappedInfo(
-        child: CircularProgressIndicator(strokeWidth: 2, color: appColors.brandColor),
+        child: LoadingSplash(),
       );
     } else if (tracks.isEmpty) {
       return ShowWrappedInfo(