+++ lib/shared/states/RecentsState.i18n.dart
@@ -1,0 +1,11 @@
+import 'package:i18n_extension/i18n_extension.dart';
+
+extension Localization on String {
+  static var _t = Translations.byLocale('en') +
+      {
+        'ru': {
+          'Error loading recent data.': 'Ошибка загрузки актуальных данных.',
+        },
+      };
+  String get i18n => localize(this, _t);
+}