+++ CHANGELOG.md
@@ -1,9 +1,35 @@
 <!--
  @since 2025.03.19, 22:43
- @changed 2025.04.15, 15:53
+ @changed 2025.06.21, 02:17
 -->
 
 # CHANGELOG
+
+## [v.0.1.12](https://github.com/lilliputten/march-tales-flutter-app/releases/tag/v.0.1.12) - 2025.06.21
+
+- Added automatic scrolling of the tracks list (issue #56).
+- Added a statistics section on the main screen.
+- Added ability to hide floating player.
+- Updated "recent" widgets: statistics moved to the bottom of the main page, styled as normal buttons.
+
+See:
+
+- [Issue #56: Implement automatic loading of tracks when scrolling](https://github.com/lilliputten/march-tales-flutter-app/issues/61)
+- [Issue #61: Add the 'Home' screen with data from recents api results](https://github.com/lilliputten/march-tales-flutter-app/issues/61)
+- [Compare with previous tag](https://github.com/lilliputten/march-tales-flutter-app/compare/march-tales-app-v.0.1.11...v.0.1.12)
+- [Release](https://github.com/lilliputten/march-tales-flutter-app/releases/tag/v.0.1.12)
+
+## [v.0.1.11](https://github.com/lilliputten/march-tales-flutter-app/releases/tag/v.0.1.11) - 2025.04.17
+
+- Added new navigation tab and page (home, 'recents').
+- Added recent data state, screen & details components with data reload feature (via `RefreshIndicator`).
+- Used standard `LoadingSplash` instead of `CircularProgressIndicator` widget wherever possible.
+
+See:
+
+- [Issue #61: Add the 'Home' screen with data from recents api results](https://github.com/lilliputten/march-tales-flutter-app/issues/61)
+- [Compare with previous tag](https://github.com/lilliputten/march-tales-flutter-app/compare/march-tales-app-v.0.1.10...v.0.1.11)
+- [Release](https://github.com/lilliputten/march-tales-flutter-app/releases/tag/v.0.1.11)
 
 ## [v.0.1.10](https://github.com/lilliputten/march-tales-flutter-app/releases/tag/v.0.1.10) - 2025.04.15